server/requirements_server.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.0.9
Werkzeug==0.11
bcrypt==2.0.0
cffi==1.3.0
cryptography==1.1
enum34==1.0.4
idna==2.0
ipaddress==1.0.14
itsdangerous==0.24
pyOpenSSL==0.15.1
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pysqlite==2.8.1
six==1.10.0
wsgiref==0.1.2
